// ui/src/generated/CaptureCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClipId } from "./ClipId";
import type { NoteId } from "./NoteId";
import type { TrackId } from "./TrackId";

export type CaptureCommand = { "type": "Capture", track: TrackId, clip: ClipId, seed_notes: NoteId, adopt_tempo: boolean, } | { "type": "Clear" } | { "type": "Status" };